Gomoku Game
Welcome to Gomoku, also known as Five in a Row! This classic board game is played on a grid where two players take turns placing their symbols ('X' and 'O') on the intersections of the lines. The objective is simple: be the first to create a continuous line of five of your symbols, either horizontally, vertically, or diagonally on the board. Strategize your moves, block your opponent, and aim for victory in this challenging battle of wits!"

Feel free to add more details or instructions to make the game more engaging for players!